Move include directories
