+++ opencog/atoms/irc-v0/IRChatStream.cc
@@ -1,5 +1,5 @@
 /*
- * opencog/atoms/sensory/IRChatStream.cc
+ * opencog/atoms/sensory-v0/IRChatStream.cc
  *
  * Copyright (C) 2020 Linas Vepstas
  * All Rights Reserved
@@ -27,7 +27,7 @@
 #include <opencog/atoms/value/StringValue.h>
 #include <opencog/atoms/value/ValueFactory.h>
 
-#include <opencog/sensory/types/atom_types.h>
+#include <opencog/sensory-v0/types/atom_types.h>
 #include "IRChatStream.h"
 
 #include "IRC.h"
