+++ opencog/atoms/terminal-v0/TerminalStream.cc
@@ -1,5 +1,5 @@
 /*
- * opencog/atoms/sensory/TerminalStream.cc
+ * opencog/atoms/sensory-v0/TerminalStream.cc
  *
  * Copyright (C) 2020 Linas Vepstas
  * All Rights Reserved
@@ -38,7 +38,7 @@
 #include <opencog/atoms/value/StringValue.h>
 #include <opencog/atoms/value/ValueFactory.h>
 
-#include <opencog/sensory/types/atom_types.h>
+#include <opencog/sensory-v0/types/atom_types.h>
 #include "TerminalStream.h"
 
 using namespace opencog;
